refactor: introduce local hitmap
- local hitmap is updated when iter
- global hitmap is updated when local hitmap drops
- metrics is updated when block is evicted from block cache
