+++ src/storage/src/monitor/mod.rs
@@ -28,6 +28,9 @@ pub use compactor_metrics::*;
 
 mod local_metrics;
 pub use local_metrics::*;
+
+mod hitmap;
+pub use hitmap::*;
 pub use risingwave_object_store::object::object_metrics::{
     ObjectStoreMetrics, GLOBAL_OBJECT_STORE_METRICS,
 };
